WIP: Implement hot reloading via jurigged
jurigged [1] allows you to update your code as it runs. It does so
by patching existing instances with new values or functions.

Jurigged does not patch all references to a given value though, so
it's important to avoid aliasing if you want changes to propagate
after reloading. This makes the `username` and related class
properties of the BaseLLMAgent problematic. I've put in a potential
solution, but we really need a cleaner story for how users configure
this stuff.

This PR does not yet address dealing with the consequences of changing
the username or the model name. To fully work, that would need to stop
the existing Flask app and register another agent. I'm still working
on whether or not this is possible.

[1] https://github.com/breuleux/jurigged