WIP: Implement hot reloading via jurigged (#11)
* WIP: Implement hot reloading via jurigged

jurigged [1] allows you to update your code as it runs. It does so
by patching existing instances with new values or functions.

Jurigged does not patch all references to a given value though, so
it's important to avoid aliasing if you want changes to propagate
after reloading. This makes the `username` and related class
properties of the BaseLLMAgent problematic. I've put in a potential
solution, but we really need a cleaner story for how users configure
this stuff.

This PR does not yet address dealing with the consequences of changing
the username or the model name. To fully work, that would need to stop
the existing Flask app and register another agent. I'm still working
on whether or not this is possible.

[1] https://github.com/breuleux/jurigged

* Make hot reloading of persona possible

* Combine watchfiles and jurigged

There's no great way to get jurigged to respond to rewire the LLM
configuration.

* Private methods don't interact perfectly with jurigged hot-loading

Problems arise when a class with a double-underscore private method
is created in a closure. jurigged handles both cases when they occur
separately, but not when they occur together, as they do in classes
created with create_llm_agent_class.

This change works around this issue by making these methods merely
nonpublic rather than private.

However, this remains a loaded footgun for developers to encounter. As
they modify the agent example, they might add their own private
functions, and this will break hot-reloading for them.

* Update base_llm_agent.py

* Update base_llm_agent.py

* updating sdk version

---------

Co-authored-by: kelmouja <kelmouja@users.noreply.github.com>